fix(rate-limiter): cast config value to int

+++ config/rate-limiter.php
@@ -6,5 +6,5 @@ return [
     /**
      * Rate limit per minute for API
      */
-    'api_rate_limit_per_minute' => env('API_RATE_LIMIT_PER_MINUTE', 60),
+    'api_rate_limit_per_minute' => (int) env('API_RATE_LIMIT_PER_MINUTE', 60),
 ];
